Made all imports follow a certain method (no longer randomly switch between import * and specific imports)
